Remove `Transaction` interface for each sub-system.
